+++ src/core/modes.ts
@@ -2,5 +2,6 @@ import './modes-utils';
 import './modes-command';
 import './modes-math';
 import './modes-text';
+import './modes-chem';
 
 export * from './modes-utils';