+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [Unreleased]
+
+### New Features
+
+-   Added support for `\phantom`, `\vphantom`, `\hphantom` and `\smash[]`
+
 ## 0.55.0 (2020-08-17)
 
 ### New Features